requirements.txt
amocrm-api==2.6.1
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Deprecated==1.2.18
distro==1.9.0
Django==5.2.3
django-celery-beat==2.8.1
django-jazzmin==3.0.1
django-solo==2.4.0
django-timezone-field==7.1
djangorestframework==3.16.0
flake8==7.2.0
gunicorn==23.0.0
idna==3.10
isort==6.0.1
kombu==5.5.4
loguru==0.7.3
mccabe==0.7.0
netaddr==1.3.0
packaging==25.0
phonenumbers==9.0.7
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.10.1
pyTelegramBotAPI==4.27.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
win32_setctime==1.2.0
wrapt==1.17.2
yookassa==3.5.0
